fix build of stratum.cpp is our codebase
there is no actual equihash stratum implementation yet, just
commented few routines, related to AuxPow and other things, which
are absent in Komodo codebase. So, now, as it built it looks more
nice and can be used as a template to continue work. this commit
is just an adoption for our codebase, further work needed.